Version 0.0.56 (dev.)
> App
- 'busboy' was transformed into a middleware.
- `.correlateAsCreateOrDelete` was renamed as `.updateThenCorrelate`.
- Promisified correlation methods in models.

> UI
- n/a

+++ config/busboy.js
@@ -1,107 +1,71 @@
 const
     fs                      = require('fs'),
     path                    = require('path'),
-    Busboy                  = require('busboy');
+    Busboy                  = require('busboy'),
+    sanitizer               = require('sanitizer');
 
 
 
 //uploader configurations
 function ImgUploadByBusboy (req, res, limits, next) {
-    // initialization
-    const notice = new ClientNoticeHandler(req);
-    const busboy = new Busboy({headers: req.headers, limits : limits});
+    const busboy = new Busboy({headers: req.headers, limits: limits});
+    const notice = [];
     const populator = {};
 
-
-    // BUSBOY-LISTENER: parse 'file' inputs
+    // BUSBOY-LISTENER: 'file' as in inputs
     busboy.on('file', (fieldName, file, fileName, encoding, MIMEType) => {
         const branch = new FileStreamBranch(fileName, MIMEType);
 
         if (branch.isAttached && branch.isAccepted) {
-            file.pipe(fs.createWriteStream(branch.fullPath).on('error', err => {    // note: error handling by listener
-                // handle all errors other than 'ENOENT'
+            file.pipe(fs.createWriteStream(branch.fullPath).on('error', err => {
                 if (err && !(err.code === 'ENOENT')) throw err;
 
-                // if sub-folder does not exist then attempt to create it
+                // if 'ENOENT': create the folder then fire the stream again
                 fs.mkdir(path.dirname(branch.fullPath), err => {
-                    if (err) throw Error(`Errors in creating ${path.dirname(branch.fullPath)}:\n${err.toString()}`);
+                    if (err) throw Error(`Errors fin creating ${path.dirname(branch.fullPath)}:\n${err.toString()}`);
                     file.pipe(fs.createWriteStream(branch.fullPath));
                 });
             }));
         } else file.resume();
 
-
-        // FILE-LISTENER: end (completed/terminated)
+        // FILE-LISTENER: after file uploaded
         file.on('end', () => {
-            const properties = _propertyReference(fieldName);
+            if (branch.isAttached && branch.isAccepted && !file.truncated) {
+                return _assignInNest(populator, _readObjectProperties(fieldName), {
+                    fileType: path.extname(branch.fileBase),
+                    fileBase: branch.fileBase,
+                    fullPath: branch.fullPath,
+                });
+            } else populator[_readObjectProperties(fieldName)[0]] = {isSkipped: true};
 
             // exception handler
-            if (!branch.isAttached || !branch.isAccepted || file.truncated) populator[properties[0]].isDiscarded = true;
-            if (!branch.isAttached) return;
-            if (!branch.isAccepted) return notice.errorUnaccepted();
-            if (file.truncated) {
-                return fs.unlink(branch.fullPath, err => {
-                    if (err) {
-                        throw new Error(`Errors in deleting: ('${branch.fullPath}', truncated):\n${err.toString()}`);
-                    } else notice.errorOversizing(limits.fileSize);
-                });
-            }
-
-            // if no exceptions
-            _updateByNestedProperty(populator, properties, {
-                fileType: path.extname(branch.fileBase),
-                fileBase: branch.fileBase,
-                fullPath: branch.fullPath,
+            if (branch.isAttached && !branch.isAccepted) return notice.push(`${fileName} is in unaccepted file types!`);
+            if (file.truncated) return fs.unlink(branch.fullPath, err => {
+                if (err) throw new Error(`Errors in deleting: ('${branch.fullPath}', truncated):\n${err.toString()}`);
+                notice.push(`${fileName} is too large! (> ${limits.limits.fileSize / 1048576} MB)`);
             });
         });
     });
 
-
-    // BUSBOY-LISTENER: parse 'field' inputs
+    // BUSBOY-LISTENER: 'field' as in inputs
     busboy.on('field', (fieldName, val) => {
-        if (val) _updateByNestedProperty(populator, _propertyReference(fieldName), req.sanitize(val));
-    });
+            if (val) _assignInNest(populator, _readObjectProperties(fieldName), sanitizer.sanitize(val));
+        });
 
-
-    // BUSBOY-LISTENER: event finished
+    // BUSBOY-LISTENER: after all streams resolved
     busboy.on('finish', () => {
-        // remove all sub-collectors (the top-level keys)
-        const mediaArray = Object.values(populator).filter(obj => obj.isDiscarded !== true);
-        if (mediaArray.length > 0) notice.infoSucceed(mediaArray.length);
-        next(null, mediaArray);
+        req.body.busboySlip = {raw: Object.values(populator).filter(obj => obj.isSkipped !== true), notice};
+        next();
     });
-
 
     // ACTIVATION: pipe busboy
     req.pipe(busboy);
 }
 
 
-// extracted functions
-function _propertyReference(source) {
-    if (!source || /[^a-zA-Z0-9._$\[\]]/g.test(source)) {
-        throw new SyntaxError('Field name cannot have special characters other than ".$[]".');
-    } else return source.match(/[a-zA-Z0-9_$]+/g);
-}
-
-function _updateByNestedProperty(obj, referenceKeys, bottomValue, index) {
-    if (!index) index = 0;
-    if (index < referenceKeys.length-1) {
-        const _extendedObj = obj[referenceKeys[index]] ? obj[referenceKeys[index]] : (obj[referenceKeys[index]] = {});
-        return _updateByNestedProperty(_extendedObj, referenceKeys, bottomValue, ++index);
-    } else obj[referenceKeys[index]] = bottomValue ? bottomValue : {};
-}
-
-
 // constructor
-function ClientNoticeHandler(req) {
-    this.errorUnaccepted = ()    => req.flash('error', `There were unaccepted file types!`);
-    this.errorOversizing = size  => req.flash('error', `Some failed in oversizing! (> ${size/1048576} MB)`);
-    this.infoSucceed     = count => req.flash('info' , `${count} File(s) successfully uploaded!`);
-}
-
 function FileStreamBranch(fileName, MIMEType) {
-    // reference
+    // structure
     this.saveTime = new Date();
     this.pathName = this.saveTime.getUTCFullYear() + `0${this.saveTime.getUTCMonth()+1}`.slice(-2);
     this.fileBase = this.saveTime.getTime() + path.extname(fileName);
@@ -114,16 +78,22 @@ function FileStreamBranch(fileName, MIMEType) {
 }
 
 
-// exports
-function ImgUploadByBusboySwitch(req, res, limits, next) {
-    if (typeof next === 'function') return ImgUploadByBusboy(req, res, limits, next);
-
-    return new Promise((resolve, reject) => {
-        ImgUploadByBusboy(req, res, limits, (err, docs) => {
-            if (err) return reject(err);
-            resolve(docs);
-        });
-    });
+// ancillary functions
+function _readObjectProperties(source) {
+    if (!source || /[^a-zA-Z0-9._$\[\]]/g.test(source)) {
+        throw new SyntaxError('Field name cannot have special characters other than ".$[]".');
+    } else return source.match(/[a-zA-Z0-9_$]+/g);
 }
 
-module.exports = ImgUploadByBusboySwitch;
+function _assignInNest(obj, referenceKeys, bottomValue, index) {
+    if (!index) index = 0;
+    if (index < referenceKeys.length-1) {
+        const _extendedObj = obj[referenceKeys[index]] ? obj[referenceKeys[index]] : (obj[referenceKeys[index]] = {});
+        return _assignInNest(_extendedObj, referenceKeys, bottomValue, ++index);
+    } else obj[referenceKeys[index]] = bottomValue ? bottomValue : {};
+}
+
+
+
+// exports
+module.exports = ImgUploadByBusboy;
